fix: add retarder losses to TC max torque request
For primary retarder (i.e. between the engine and the torque converter),
the retarder losses are not added to the dryRun max torque request
inputs for torque converter component, which lead to engine overload in
the final response when the max torque is reached.

The fix add the Retarder torque losses to the Databus for them to be used
in the TorqueConverter request. If the Retarder is not primary, the value
read will be null (as in non-primary configuration, the request to the
retarder will be made after the request from the TorqueConverter), and
not taken into account.

Tested on all article 10 issues submitted till 02/10/2025 with no effect
on consumption, except for one issue concerning an HEV, and only mission
(LongHaul LowLoading, -0.25% FC in this case).

closes: #814, #990, #1057