fix: DefaultDriverStrategy - add accel action after gearshift + roll + brake if overload
In #702 on after a gearshift with negative gradient,
- driver goes to roll leading to overspeed because it does not realize
the target acceleration and look for another positive one
overspeed is triggered due to a condition with 0 tolerance.
- driver then goes to brake leading to an overload
- testing an accelerate action after works because with a response similar to
Rolling, except it does not trigger the  overspeed condition and falls into
the torque tolerance for Underload

Other option: add a tolerance to the detection of overspeed in Driver.cs
line 502

Tested on all article 10 issues submitted till 25/10/2025, without any
change of fuel consumption on other unrelated issues.

Solves #657, #702